Base Recon Mission - Complete
Changes made:

- Mi-8 has been replaced with a Ka-60 (because the orca kept crashing
for some reason, either way. it looks the same and sounds the same and
will only be there about 2 mins. there's no unarmed variant
unfortunately, but it's under strict order to not shoot anything)

- AI bugs at base neutralised (i.e. the weird thing they were doing
where they just up and decided to start running around is not happening,
because they're stuck in animations)

- Intel has been added, and i think it works pretty well. (THANKS
@jonpas). Someone has to search a cabinet, there's a progress bar for
taking a scan of a document, that scan then appears as a diary entry
under PDA and the player receives a hint to tell them. The intel is
encrypted/enciphered (I have the key) so someone can either attempt to
crack it (omg ARG) or i can just mention in the mid-week SMEAC that
we've received the key/a large amount of them, and that we can decipher
when the intel is recovered.

also included is the .sqf file for the intel search and diary entry
(again, thanks @jonpas )